+++ src/mynotes.sh
@@ -1,24 +1,34 @@
 #!/bin/bash
 
+#######################################################################################################################
+# This script is a controlling script for the mynotes.py server.  With this script you can:
+#   1. start the mynotes server: mynotes start
+#   2. stop the mynotes server: mynote stop
+#   3. schedule a note with the mynotes server: mynote "text..."
+#######################################################################################################################
+
 [ -z "$1" ] && echo "Usage: mynotes [start|stop|text]" && exit 0
 
 declare mydir="$HOME/mynotes"
 declare log="$HOME/logs/mynotes.log"
 
+# Initalize .settings is the is a first run
 if [ ! -d $HOME/mynotes ]
 then
     mkdir $HOME/mynotes
     echo "file: 1" > $mydir/.settings
 fi
 
-
 declare option="$1"
-declare -i file_idx=$(grep "^file: " .settings | awk '{print $2}')
+declare -i file_idx=$(grep "^file: " $mydir/.settings | awk '{print $2}')
 declare file_name="$file_idx.txt"
 
+# Increment the file name index value
 echo "file: $(($file_idx+1))" > $mydir/.settings
 
+#######################################################################################################################
 # Function to check and format the $sched variable
+#######################################################################################################################
 valid_sched_format() {
     local sched="$1"
     local current_date
@@ -39,10 +49,15 @@ valid_sched_format() {
     fi
 }
 
+#######################################################################################################################
+#######################################################################################################################
+# Execute the user's option
+#######################################################################################################################
+#######################################################################################################################
 case "$option" in
     "start")
         touch /tmp/.mynotes.running
-        $HOME/bin/mynotes.py "$mydir" >> $log
+        python -u $HOME/bin/mynotes.py "$mydir" >> $log &
         ;;
     
     "stop")
@@ -50,25 +65,31 @@ case "$option" in
         ;;
     
     *)
-        sched=""
-        text="$1"
-
-        if [ -z "$2" ]
+        if [ -f /tmp/.mynotes.running ]
         then
-            echo "Enter schedule (yyyy-mm-dd hh:mm:ss or hh:mm:ss)"
-            read sched
+            sched=""
+            text="$1"
+
+            if [ -z "$2" ]
+            then
+                echo "Enter schedule (yyyy-mm-dd hh:mm:ss or hh:mm:ss)"
+                read sched
+            else
+                sched="$2"
+            fi
+
+            # If it's a valid date, this function will echo it back or nother
+            # If the date is missing, it will plug it in with the current one
+            clean_sched=$(valid_sched_format "$sched")
+
+            if [ ! -z "$clean_sched" ]
+            then
+                # Schedule work for the mynotes server
+                printf "%s\n%s" "$clean_sched" "$text" > "$mydir/$file_name"
+            else
+                echo "ERROR: $sched is not in form yyyy-mm-dd hh:mm:ss or hh:mm:ss"
+            fi
         else
-            sched="$2"
+            echo "ERROR: mynotes server not running"
         fi
-
-        clean_sched=$(valid_sched_format "$sched")
-
-        if [ ! -z "$clean_sched" ]
-        then
-            printf "%s\n%s" "$clean_sched" "$text" > "$mydir/$file_name"
-        else
-            echo "ERROR: $sched is not in form yyyy-mm-dd hh:mm:ss or hh:mm:ss"
-        fi
-        ;;
-
 esac
